LPS-73481 Move upgrade processes implemented after the release of GA3 to a new upgrade version
