+++ portal-impl/src/com/liferay/portal/upgrade/UpgradeProcess_7_0_4.java
@@ -16,33 +16,32 @@ package com.liferay.portal.upgrade;
 
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
 import com.liferay.portal.kernel.util.ReleaseInfo;
-import com.liferay.portal.upgrade.v7_0_3.UpgradeGroup;
-import com.liferay.portal.upgrade.v7_0_3.UpgradeMessageBoards;
-import com.liferay.portal.upgrade.v7_0_3.UpgradeModules;
-import com.liferay.portal.upgrade.v7_0_3.UpgradeOracle;
-import com.liferay.portal.upgrade.v7_0_3.UpgradeOrganization;
-import com.liferay.portal.upgrade.v7_0_3.UpgradeSQLServer;
-import com.liferay.portal.upgrade.v7_0_3.UpgradeSybase;
+import com.liferay.portal.upgrade.v7_0_4.UpgradeContact;
+import com.liferay.portal.upgrade.v7_0_4.UpgradeEmailAddress;
+import com.liferay.portal.upgrade.v7_0_4.UpgradeExpando;
+import com.liferay.portal.upgrade.v7_0_4.UpgradeMBMailingList;
+import com.liferay.portal.upgrade.v7_0_4.UpgradeSchema;
+import com.liferay.portal.upgrade.v7_0_4.UpgradeUser;
 
 /**
- * @author Adolfo Pérez
+ * @author Samuel Ziemer
  */
-public class UpgradeProcess_7_0_3 extends UpgradeProcess {
+public class UpgradeProcess_7_0_4 extends UpgradeProcess {
 
 	@Override
 	public int getThreshold() {
-		return ReleaseInfo.RELEASE_7_0_3_BUILD_NUMBER;
+		return ReleaseInfo.RELEASE_7_0_4_BUILD_NUMBER;
 	}
 
 	@Override
 	protected void doUpgrade() throws Exception {
-		upgrade(UpgradeGroup.class);
-		upgrade(UpgradeMessageBoards.class);
-		upgrade(UpgradeModules.class);
-		upgrade(UpgradeOrganization.class);
-		upgrade(UpgradeOracle.class);
-		upgrade(UpgradeSQLServer.class);
-		upgrade(UpgradeSybase.class);
+		upgrade(UpgradeSchema.class);
+
+		upgrade(UpgradeContact.class);
+		upgrade(UpgradeEmailAddress.class);
+		upgrade(UpgradeExpando.class);
+		upgrade(UpgradeMBMailingList.class);
+		upgrade(UpgradeUser.class);
 
 		clearIndexesCache();
 	}
