+++ portal-impl/src/com/liferay/portal/upgrade/v7_0_4/UpgradeMBMailingList.java
@@ -12,21 +12,22 @@
  * details.
  */
 
-package com.liferay.portal.upgrade.v7_0_3;
+package com.liferay.portal.upgrade.v7_0_4;
 
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
-import com.liferay.portal.upgrade.v7_0_3.util.ContactTable;
+import com.liferay.portal.upgrade.v7_0_4.util.MBMailingListTable;
 
 /**
  * @author Ugurcan Cetin
  */
-public class UpgradeContact extends UpgradeProcess {
+public class UpgradeMBMailingList extends UpgradeProcess {
 
 	@Override
 	protected void doUpgrade() throws Exception {
 		alter(
-			ContactTable.class,
-			new AlterColumnType("emailAddress", "VARCHAR(254) null"));
+			MBMailingListTable.class,
+			new AlterColumnType("emailAddress", "VARCHAR(254) null"),
+			new AlterColumnType("outEmailAddress", "VARCHAR(254) null"));
 	}
 
 }
