+++ portal-impl/src/portal.properties
@@ -132,7 +132,8 @@
         com.liferay.portal.upgrade.UpgradeProcess_6_2_0\,\
         com.liferay.portal.upgrade.UpgradeProcess_7_0_0\,\
         com.liferay.portal.upgrade.UpgradeProcess_7_0_1\,\
-        com.liferay.portal.upgrade.UpgradeProcess_7_0_3
+        com.liferay.portal.upgrade.UpgradeProcess_7_0_3\,\
+        com.liferay.portal.upgrade.UpgradeProcess_7_0_4
 
     upgrade.processes.6012=${upgrade.processes.master}
     upgrade.processes.6100=${upgrade.processes.master}
